Split out parts of `re_data_source` to `re_data_loader`
This makes the dependency list of re_sdk much lighter
