+++ Cargo.toml
@@ -32,6 +32,7 @@ re_analytics = { path = "crates/re_analytics", version = "=0.16.0-alpha.2", defa
 re_build_info = { path = "crates/re_build_info", version = "=0.16.0-alpha.2", default-features = false }
 re_build_tools = { path = "crates/re_build_tools", version = "=0.16.0-alpha.2", default-features = false }
 re_crash_handler = { path = "crates/re_crash_handler", version = "=0.16.0-alpha.2", default-features = false }
+re_data_loader = { path = "crates/re_data_loader", version = "=0.16.0-alpha.2", default-features = false }
 re_data_source = { path = "crates/re_data_source", version = "=0.16.0-alpha.2", default-features = false }
 re_data_store = { path = "crates/re_data_store", version = "=0.16.0-alpha.2", default-features = false }
 re_data_ui = { path = "crates/re_data_ui", version = "=0.16.0-alpha.2", default-features = false }
