+++ crates/re_data_loader/README.md
@@ -1,0 +1,8 @@
+# re_data_loader
+
+Part of the [`rerun`](https://github.com/rerun-io/rerun) family of crates.
+
+[![Latest version](https://img.shields.io/crates/v/re_data_loader.svg)](https://crates.io/crates/re_data_loader)
+[![Documentation](https://docs.rs/re_data_loader/badge.svg)](https://docs.rs/re_data_loader)
+![MIT](https://img.shields.io/badge/license-MIT-blue.svg)
+![Apache](https://img.shields.io/badge/license-Apache-blue.svg)
