+++ crates/re_data_loader/src/loader_archetype.rs
