+++ crates/re_data_loader/src/loader_directory.rs
