+++ crates/re_sdk/Cargo.toml
@@ -28,7 +28,7 @@ default = []
 ##
 ## See our `log_file` example and <https://www.rerun.io/docs/reference/data-loaders/overview>
 ## for more information.
-data_loaders = ["dep:re_data_source", "dep:re_smart_channel"]
+data_loaders = ["dep:re_data_loader", "dep:re_smart_channel"]
 
 ## Support serving a web viewer over HTTP.
 ##
@@ -68,7 +68,7 @@ thiserror.workspace = true
 
 # Optional dependencies
 
-re_data_source = { workspace = true, optional = true }
+re_data_loader = { workspace = true, optional = true }
 re_smart_channel = { workspace = true, optional = true }
 re_ws_comms = { workspace = true, optional = true }
 re_web_viewer_server = { workspace = true, optional = true }
