+++ crates/re_sdk/src/lib.rs
@@ -87,7 +87,7 @@ pub use re_types_core::{
 };
 
 #[cfg(feature = "data_loaders")]
-pub use re_data_source::{DataLoader, DataLoaderError, DataLoaderSettings, LoadedData};
+pub use re_data_loader::{DataLoader, DataLoaderError, DataLoaderSettings, LoadedData};
 
 /// Methods for spawning the web viewer and streaming the SDK log stream to it.
 #[cfg(feature = "web_viewer")]
@@ -103,7 +103,7 @@ pub mod external {
     pub use re_log_types::external::*;
 
     #[cfg(feature = "data_loaders")]
-    pub use re_data_source;
+    pub use re_data_loader;
 
     #[cfg(feature = "log")]
     pub use log;
