+++ crates/re_sdk/src/recording_stream.rs
@@ -83,10 +83,10 @@ pub enum RecordingStreamError {
     #[error(transparent)]
     DataReadError(#[from] re_log_types::DataReadError),
 
-    /// An error occurred while attempting to use a [`re_data_source::DataLoader`].
+    /// An error occurred while attempting to use a [`re_data_loader::DataLoader`].
     #[cfg(feature = "data_loaders")]
     #[error(transparent)]
-    DataLoaderError(#[from] re_data_source::DataLoaderError),
+    DataLoaderError(#[from] re_data_loader::DataLoaderError),
 }
 
 /// Results that can occur when creating/manipulating a [`RecordingStream`].
@@ -1062,11 +1062,11 @@ impl RecordingStream {
         Ok(())
     }
 
-    /// Logs the file at the given `path` using all [`re_data_source::DataLoader`]s available.
+    /// Logs the file at the given `path` using all [`re_data_loader::DataLoader`]s available.
     ///
     /// A single `path` might be handled by more than one loader.
     ///
-    /// This method blocks until either at least one [`re_data_source::DataLoader`] starts
+    /// This method blocks until either at least one [`re_data_loader::DataLoader`] starts
     /// streaming data in or all of them fail.
     ///
     /// See <https://www.rerun.io/docs/reference/data-loaders/overview> for more information.
@@ -1080,11 +1080,11 @@ impl RecordingStream {
         self.log_file(filepath, None, entity_path_prefix, static_)
     }
 
-    /// Logs the given `contents` using all [`re_data_source::DataLoader`]s available.
+    /// Logs the given `contents` using all [`re_data_loader::DataLoader`]s available.
     ///
     /// A single `path` might be handled by more than one loader.
     ///
-    /// This method blocks until either at least one [`re_data_source::DataLoader`] starts
+    /// This method blocks until either at least one [`re_data_loader::DataLoader`] starts
     /// streaming data in or all of them fail.
     ///
     /// See <https://www.rerun.io/docs/reference/data-loaders/overview> for more information.
@@ -1145,7 +1145,7 @@ impl RecordingStream {
         };
 
         if let Some(contents) = contents {
-            re_data_source::load_from_file_contents(
+            re_data_loader::load_from_file_contents(
                 &settings,
                 re_log_types::FileSource::Sdk,
                 filepath,
@@ -1153,7 +1153,7 @@ impl RecordingStream {
                 &tx,
             )?;
         } else {
-            re_data_source::load_from_path(
+            re_data_loader::load_from_path(
                 &settings,
                 re_log_types::FileSource::Sdk,
                 filepath,
