+++ crates/re_viewer/Cargo.toml
@@ -39,6 +39,7 @@ analytics = ["dep:re_analytics"]
 [dependencies]
 # Internal:
 re_build_info.workspace = true
+re_data_loader.workspace = true
 re_data_source.workspace = true
 re_data_store.workspace = true
 re_data_ui.workspace = true
