+++ examples/rust/custom_data_loader/src/main.rs
@@ -1,4 +1,4 @@
-//! This example demonstrates how to implement and register a [`re_data_source::DataLoader`] into
+//! This example demonstrates how to implement and register a [`re_data_loader::DataLoader`] into
 //! the Rerun Viewer in order to add support for loading arbitrary files.
 //!
 //! Usage:
@@ -7,7 +7,7 @@
 //! ```
 
 use rerun::{
-    external::{anyhow, re_build_info, re_data_source, re_log},
+    external::{anyhow, re_build_info, re_data_loader, re_log},
     log::{DataRow, RowId},
     EntityPath, TimePoint,
 };
@@ -15,7 +15,7 @@ use rerun::{
 fn main() -> anyhow::Result<std::process::ExitCode> {
     re_log::setup_logging();
 
-    re_data_source::register_custom_data_loader(HashLoader);
+    re_data_loader::register_custom_data_loader(HashLoader);
 
     let build_info = re_build_info::build_info!();
     rerun::run(build_info, rerun::CallSource::Cli, std::env::args())
@@ -24,43 +24,43 @@ fn main() -> anyhow::Result<std::process::ExitCode> {
 
 // ---
 
-/// A custom [`re_data_source::DataLoader`] that logs the hash of file as a [`rerun::TextDocument`].
+/// A custom [`re_data_loader::DataLoader`] that logs the hash of file as a [`rerun::TextDocument`].
 struct HashLoader;
 
-impl re_data_source::DataLoader for HashLoader {
+impl re_data_loader::DataLoader for HashLoader {
     fn name(&self) -> String {
         "rerun.data_loaders.HashLoader".into()
     }
 
     fn load_from_path(
         &self,
-        _settings: &rerun::external::re_data_source::DataLoaderSettings,
+        _settings: &rerun::external::re_data_loader::DataLoaderSettings,
         path: std::path::PathBuf,
-        tx: std::sync::mpsc::Sender<re_data_source::LoadedData>,
-    ) -> Result<(), re_data_source::DataLoaderError> {
+        tx: std::sync::mpsc::Sender<re_data_loader::LoadedData>,
+    ) -> Result<(), re_data_loader::DataLoaderError> {
         let contents = std::fs::read(&path)?;
         if path.is_dir() {
-            return Err(re_data_source::DataLoaderError::Incompatible(path)); // simply not interested
+            return Err(re_data_loader::DataLoaderError::Incompatible(path)); // simply not interested
         }
         hash_and_log(&tx, &path, &contents)
     }
 
     fn load_from_file_contents(
         &self,
-        _settings: &rerun::external::re_data_source::DataLoaderSettings,
+        _settings: &rerun::external::re_data_loader::DataLoaderSettings,
         filepath: std::path::PathBuf,
         contents: std::borrow::Cow<'_, [u8]>,
-        tx: std::sync::mpsc::Sender<re_data_source::LoadedData>,
-    ) -> Result<(), re_data_source::DataLoaderError> {
+        tx: std::sync::mpsc::Sender<re_data_loader::LoadedData>,
+    ) -> Result<(), re_data_loader::DataLoaderError> {
         hash_and_log(&tx, &filepath, &contents)
     }
 }
 
 fn hash_and_log(
-    tx: &std::sync::mpsc::Sender<re_data_source::LoadedData>,
+    tx: &std::sync::mpsc::Sender<re_data_loader::LoadedData>,
     filepath: &std::path::Path,
     contents: &[u8],
-) -> Result<(), re_data_source::DataLoaderError> {
+) -> Result<(), re_data_loader::DataLoaderError> {
     use std::collections::hash_map::DefaultHasher;
     use std::hash::{Hash, Hasher};
